requirements.txt
aioredis==1.1.0
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
certifi==2018.4.16
channels==2.1.2
channels-redis==2.2.1
constantly==15.1.0
daphne==2.2.0
dj-database-url==0.5.0
Django==2.0.6
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.1.5
gunicorn==19.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
pipenv==2018.7.1
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
zope.interface==4.5.0
